moved guard out to its own file
