+++ docs/_modules/syft.core.hooks.torch.rst
@@ -1,0 +1,30 @@
+syft\.core\.hooks\.torch package
+================================
+
+Submodules
+----------
+
+syft\.core\.hooks\.torch\.guard module
+--------------------------------------
+
+.. automodule:: syft.core.hooks.torch.guard
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+syft\.core\.hooks\.torch\.hook module
+-------------------------------------
+
+.. automodule:: syft.core.hooks.torch.hook
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: syft.core.hooks.torch
+    :members:
+    :undoc-members:
+    :show-inheritance:
