+++ syft/core/hooks/torch/__init__.py
@@ -1,0 +1,2 @@
+from . import guard
+from .hook import TorchHook
