+++ syft/core/hooks/torch/hook.py
@@ -5,9 +5,10 @@ import json
 import types
 import functools
 import importlib
-from .. import workers
-from .. import utils
-from .base import BaseHook
+from ... import workers
+from ... import utils
+from ..base import BaseHook
+from .guard import TorchGuard
 
 
 class TorchHook(BaseHook):
@@ -130,23 +131,7 @@ class TorchHook(BaseHook):
         # Torch functions we don't want to override
         self.torch_exclude = ['save', 'load', 'typename']
 
-        # Safety checks for serializing and deserializing torch objects
-        # Desperately needs stress testing before going out in the wild
-        self.map_tensor_type = {
-            'torch.FloatTensor': torch.FloatTensor,
-            'torch.DoubleTensor': torch.DoubleTensor,
-            'torch.HalfTensor': torch.HalfTensor,
-            'torch.ByteTensor': torch.ByteTensor,
-            'torch.CharTensor': torch.CharTensor,
-            'torch.ShortTensor': torch.ShortTensor,
-            'torch.IntTensor': torch.IntTensor,
-            'torch.LongTensor': torch.LongTensor
-        }
-        self.map_var_type = {
-            'torch.autograd.variable.Variable': torch.autograd.variable.Variable,
-            'torch.nn.parameter.Parameter': torch.nn.parameter.Parameter
-        }
-        self.map_torch_type = dict(self.map_tensor_type, **self.map_var_type)
+        self.guard = TorchGuard()
 
         if(not hasattr(torch, 'hooked')):
             if(verbose):
@@ -164,53 +149,6 @@ class TorchHook(BaseHook):
                 print("WARNING: Torch seems to be already overloaded... skipping...")
 
     # ######## BEGIN PUBLIC METHODS #########
-
-    def types_guard(self, torch_type_str):
-        """types_guard(torch_type_str) -> torch.Tensor or torch.autograd.Variable
-
-        This method converts strings into a type reference. This prevents
-        deserialized JSON from being able to instantiate objects of arbitrary
-        type which would be a security concern.
-
-        :Parameters:
-
-        * **torch_type_str (string)** A string representing the type of object that is
-          to be returned.
-
-        * **out (a torch type)** The type the string refersto (if it's present in the
-          acceptible list self.map_torch_type)
-
-        :Example:
-
-        >>> from syft.core.hooks import TorchHook
-        >>> hook = TorchHook()
-        Hooking into Torch...
-        Overloading Complete.
-        >>> torch_type = hook.types_guard('torch.FloatTensor')
-        >>> x = torch_type([1,2,3,4,5])
-        >>> x
-         1
-         2
-         3
-         4
-         5
-        [torch.FloatTensor of size 5]
-        """
-        try:
-            return self.map_torch_type[torch_type_str]
-        except KeyError:
-            raise TypeError(
-                "Tried to receive a non-Torch object of type {}.".format(
-                    torch_type_str))
-
-    def tensor_contents_guard(self, contents):
-        """tensor_contents_guard(contents) -> contents
-        TODO: check to make sure the incoming list isn't dangerous to use for
-               constructing a tensor (likely non-trivial). Accepts the list of JSON objects
-               and returns the list of JSON ojects. Should throw and exception if there's a
-               security concern.
-        """
-        return contents
 
     @staticmethod
     def get_owners(tensorvars):
@@ -488,7 +426,7 @@ class TorchHook(BaseHook):
         """
         # TODO: extend to iterables of tensor pointers
         try:
-            torch_type = self.map_torch_type[torch_type]
+            torch_type = self.guard.types_guard(torch_type)
         except KeyError:
             raise TypeError(
                 "Tried to receive a non-Torch object of type {}.".format(
@@ -761,7 +699,7 @@ class TorchHook(BaseHook):
         def deser(self, obj_msg):
 
             # this could be a significant failure point, security-wise
-            data = hook_self.tensor_contents_guard(obj_msg['data'])
+            data = hook_self.guard.tensor_contents_guard(obj_msg['data'])
             v = self(data)
             return v
 
@@ -999,7 +937,7 @@ class TorchHook(BaseHook):
 
             if 'data' in obj_msg.keys():
                 data_msg = json.loads(obj_msg['data'])
-                tensor_type = hook_self.types_guard(data_msg['torch_type'])
+                tensor_type = hook_self.guard.types_guard(data_msg['torch_type'])
                 data_obj = tensor_type.deser(tensor_type, data_msg)
                 # data_obj = hook_self.build_tensor(data_msg, tensor_type)
                 data = hook_self.local_worker.handle_register(
@@ -1010,7 +948,7 @@ class TorchHook(BaseHook):
 
                     grad_msg = json.loads(obj_msg['grad'])
 
-                    var_type = hook_self.types_guard(grad_msg['torch_type'])
+                    var_type = hook_self.guard.types_guard(grad_msg['torch_type'])
                     grad_obj = hook_self._build_var(grad_msg, var_type)
 
                     grad = hook_self.local_worker.handle_register(grad_obj, grad_msg,
@@ -1057,7 +995,7 @@ class TorchHook(BaseHook):
 
         if 'data' in obj_msg.keys():
             data_msg = json.loads(obj_msg['data'])
-            tensor_type = self.types_guard(data_msg['torch_type'])
+            tensor_type = self.guard.types_guard(data_msg['torch_type'])
             data_obj = tensor_type.deser(tensor_type, data_msg)
             # data_obj = self.build_tensor(data_msg, tensor_type)
             data = self.local_worker.handle_register(
@@ -1066,7 +1004,7 @@ class TorchHook(BaseHook):
         if 'grad' in obj_msg.keys():
             if obj_msg['grad'] is not None:
                 grad_msg = json.loads(obj_msg['grad'])
-                var_type = self.types_guard(grad_msg['torch_type'])
+                var_type = self.guard.types_guard(grad_msg['torch_type'])
                 grad_obj = self._build_var(grad_msg, var_type)
                 grad = self.local_worker.handle_register(
                     grad_obj, grad_msg, temporary=True)
